build: ContractClient-related things in one dir
